+++ src/Common/Commands.Common.Authentication.Test/Properties/AssemblyInfo.cs
@@ -28,6 +28,6 @@ using Xunit;
 
 [assembly: ComVisible(false)]
 [assembly: CLSCompliant(false)]
-[assembly: AssemblyVersion("1.0.4")]
-[assembly: AssemblyFileVersion("1.0.4")]
+[assembly: AssemblyVersion("1.3.2")]
+[assembly: AssemblyFileVersion("1.3.2")]
 [assembly: CollectionBehavior(DisableTestParallelization = true)]