+++ src/ServiceManagement/Services/Commands/ServiceBus/GetAzureSBNamespace.cs
@@ -35,7 +35,7 @@ namespace Microsoft.WindowsAzure.Commands.ServiceBus
         /// </summary>
         public override void ExecuteCmdlet()
         {
-            Client = Client ?? new ServiceBusClientExtensions(Profile, Profile.Context.Subscription);
+            Client = Client ?? new ServiceBusClientExtensions(Profile);
 
             if (string.IsNullOrEmpty(Name))
             {