+++ ChangeLog.md
@@ -1,4 +1,50 @@
-﻿##2016.03.30 version 1.2.3
+##2016.05.03 version 1.4.0
+* Azure Storage
+  * Upgrade to Microsoft.Azure.Management.Storage nuget package v5.0 preview
+  * Add Encryption and Hot/Cool features support to resource mode storage account cmdlets
+    - New-AzureRmStorageAccount
+    - Set-AzureRmStorageAccount
+  * Add "Add" and "Create" permission to Blob SAS cmdlets
+    - New-AzureStorageBlobSASToken
+    - New-AzureStorageContainerSASToken
+    - New-AzureStorageContainerStoredAccessPolicy
+    - Set-AzureStorageContainerStoredAccessPolicy
+  * Add "Create" permission to File SAS cmdlets
+    - New-AzureStorageFileSASToken
+    - New-AzureStorageShareSASToken
+    - New-AzureStorageShareStoredAccessPolicy
+    - Set-AzureStorageShareStoredAccessPolicy
+
+##2016.04.19 version 1.3.2
+* Add support for specifying NIC/VMSS as application gateway backend address
+* Fix HDI ADL cluster creation and live test
+* Fix WAPack cmdlet Proxy issue for WAP
+* Fix Dynamic Memory Issue while setting VM
+* Update Azure Gov STS and Traffic Manager
+* Compute
+  - Upgrade to Microsoft.Azure.Management.Compute nuget package v13.0 
+* HDInsignt
+  - Upgrade to Microsoft.Azure.Management.HDInsight nuget package v1.0.14
+* Resource Manager 
+  - Change api-version for Policy and Locks operations
+  - Change api version for Deployments operations
+* Web Apps
+  - Add backup and restore cmdlets
+    - Restore-AzureWebApp
+    - Edit-AzureRmWebAppBackupConfiguration
+    - Get-AzureRmWebAppBackupConfiguration
+    - Get-AzureRmWebAppBackup
+    - New-AzureRmWebAppBackup
+    - Remove-AzureRmWebAppBackup
+    - Restore-AzureRmWebAppBackup
+    - Get-AzureRmWebAppBackupList
+    - New-AzureRmWebAppDatabaseBackupSetting
+  - Upgrade to Microsoft.WindowsAzure.Management.WebSite nuget package v5.0
+
+##2016.03.30 version 1.3.0
+* AzureRM module
+  - Installation performance fix
+* New Azure CDN cmdlets
 * Azure Storage
   * Made Protocol parameter in following cmdlets to be nullable and optional
     - New-AzureStorageBlobSASToken
@@ -8,6 +54,10 @@
     - New-AzureStorageQueueSASToken
     - New-AzureStorageTableSASToken
     - New-AzureStorageAccountSASToken
+* Add Export-AzureRmResourceGroupTemplate cmdlet
+* Add VirtualMachineScaleSet cmdlet
+* Add "EA" value for permission parameter of Set-AzurePlatformVMImage
+* Documentation improvements in cmdlets 
 
 ##2016.03.03 version 1.2.2
 * Azure Compute (ARM):